examples/langgraph_agent/requirements.txt
-e ../..
langgraph
python-dotenv